requirements.txt
dj-rest-auth==5.0.2
Django==4.2.11
django-allauth==0.61.1
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-import-export==3.3.7
django-ratelimit==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
openpyxl==3.1.2
pillow==10.2.0
psycopg-binary==3.1.18
psycopg2==2.9.9
pycparser==2.21
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
transliterate==1.10.2
